refactor(cli): extract agents file diff controller
